Re-do middleware implementation
This hotfix takes the middleware implementation utilized in my discourse-windowslive-auth plugin and implements it here. This fixes the issues of settings updates not being loaded unless the Discourse container was relaunched.

This also adds the enabled_site_setting option as described in the documentation for Discourse Plugins, making the plugin more in-line with official specifications.